Add files via upload
Uploading existing work.

This simple script allows, at the press of a hotkey, the creation of folders in statically set locations in windows, the creation of a folder in outlook, and the creation of two rules to move messages to that new folder.

Unfortunately, the way it accomplishes the creation of rules is rather... jank, relying on Outlook's built-in Visual Basic engine, and the creation of a new menu button in outlook. This would not be necessary save for a type related bug affecting the ability to assign a rule based off subject line text in Outlooks API. I know this to be an API level issue as it is recreatable when accessing outlook both via the AHK library used here, and via Powershell. Namely, textrule condtions in the Outlook object model have no valid acceptable type for their string data when assigned to a subject rule, resulting in typeerrors across the board. For more information on the TextRule condition, see

https://docs.microsoft.com/en-us/office/vba/api/outlook.textrulecondition